+++ src/github.com/stellar/horizon/action.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/stellar/horizon/actions"
-	"github.com/stellar/horizon/db2"
 	"github.com/stellar/horizon/db2/core"
 	"github.com/stellar/horizon/db2/history"
 	"github.com/stellar/horizon/httpx"
@@ -39,43 +38,6 @@ func (action *Action) CoreQ() *core.Q {
 	}
 
 	return action.cq
-}
-
-// GetPagingParams modifies the base GetPagingParams method to replace
-// cursors that are "now" with the last seen ledger's cursor.
-func (action *Action) GetPagingParams() (cursor string, order string, limit uint64) {
-	if action.Err != nil {
-		return
-	}
-
-	cursor, order, limit = action.Base.GetPagingParams()
-
-	if cursor == "now" {
-		tid := toid.ID{
-			LedgerSequence:   ledger.CurrentState().HistoryLatest,
-			TransactionOrder: toid.TransactionMask,
-			OperationOrder:   toid.OperationMask,
-		}
-		cursor = tid.String()
-	}
-
-	return
-}
-
-// GetPageQuery is a helper that returns a new db.PageQuery struct initialized
-// using the results from a call to GetPagingParams()
-func (action *Action) GetPageQuery() db2.PageQuery {
-	if action.Err != nil {
-		return db2.PageQuery{}
-	}
-
-	r, err := db2.NewPageQuery(action.GetPagingParams())
-
-	if err != nil {
-		action.Err = err
-	}
-
-	return r
 }
 
 // HistoryQ provides access to queries that access the history portion of