+++ CHANGELOG.md
@@ -6,6 +6,12 @@ file.  This project adheres to [Semantic Versioning](http://semver.org/).
 As this project is pre 1.0, breaking changes may happen for minor version
 bumps.  A breaking change will get clearly notified in this log.
 
+## [Unreleased]
+
+### Changed
+
+- BREAKING CHANGE: The payments endpoint now includes `account_merge` operations in the response.
+
 ## [v0.10.1] - 2017-03-29
 
 ### Fixed